Re-structured JS and added multi-model parameter settings
